+++ otelconf/log.go
@@ -105,10 +105,6 @@ func logExporter(ctx context.Context, exporter LogRecordExporter) (sdklog.Export
 			return otlpGRPCLogExporter(ctx, exporter.OTLPGrpc)
 		}
 	}
-	if exporter.OTLPFileDevelopment != nil {
-		// TODO: implement file exporter https://github.com/open-telemetry/opentelemetry-go/issues/5408
-		return nil, newErrInvalid("otlp_file/development")
-	}
 
 	if exportersConfigured > 1 {
 		return nil, newErrInvalid("must not specify multiple exporters")