+++ otelconf/resource.go
@@ -12,25 +12,6 @@ import (
 	"go.opentelemetry.io/contrib/otelconf/internal/kv"
 )
 
-func resourceOpts(detectors []ExperimentalResourceDetector) []resource.Option {
-	opts := []resource.Option{}
-	for _, d := range detectors {
-		if d.Container != nil {
-			opts = append(opts, resource.WithContainer())
-		}
-		if d.Host != nil {
-			opts = append(opts, resource.WithHost(), resource.WithOS())
-		}
-		if d.Process != nil {
-			opts = append(opts, resource.WithProcess())
-		}
-		if d.Service != nil {
-			opts = append(opts, resource.WithService())
-		}
-	}
-	return opts
-}
-
 func newResource(r *Resource) (*resource.Resource, error) {
 	if r == nil {
 		return resource.Default(), nil
@@ -48,10 +29,6 @@ func newResource(r *Resource) (*resource.Resource, error) {
 	opts := []resource.Option{
 		resource.WithAttributes(attrs...),
 		resource.WithSchemaURL(schema),
-	}
-
-	if r.DetectionDevelopment != nil {
-		opts = append(opts, resourceOpts(r.DetectionDevelopment.Detectors)...)
 	}
 
 	return resource.New(context.Background(), opts...)